Add DefaultErrorView and reorganize DynamicList structure for improved maintainability. Introduced separate files for previews and models, enhancing clarity and usability. Updated documentation to reflect new file structure and components.
